Some code review changes
* Revert line endings change in Python.Runtime.csproj
* Split maybe serialize into respective class files
* Name changes for consistency
